Ignore array and object trailing commas when reading settings
Windows Terminal uses the JSONC format for its settings, but we are parsing them with a standard JSON parser. To work around this, we use JsMin to remove all comments. Trailing commas however can still trip the parser.

With this commit, we try to remove trailing commas with string replacement. It's not perfect, the proper fix would be to find a JSONC parser.